:recycle: refactor: change all ts style to css
