+++ src/components/Badge/Badge.css
@@ -1,0 +1,10 @@
+.badge-container {
+  color: white;
+  background-color: rgb(37 99 235);
+  font-weight: 500;
+  text-transform: uppercase;
+  padding: 0.25rem 0.5rem;
+  line-height: 1rem;
+  font-size: 0.75rem;
+  border-radius: 999px;
+}
