--- src/components/Badge/BadgeStyle.ts
@@ -1,14 +1,0 @@
-import { css } from "lit";
-
-export default css`
-  .badge-container {
-    color: white;
-    background-color: rgb(37 99 235);
-    font-weight: 500;
-    text-transform: uppercase;
-    padding: 0.25rem 0.5rem;
-    line-height: 1rem;
-    font-size: 0.75rem;
-    border-radius: 999px;
-  }
-`;
