+++ src/components/Divider/Divider.css
@@ -1,0 +1,8 @@
+.divider-container {
+  padding: 0.5rem;
+}
+
+.divider-container div {
+  display: flex;
+  border-top: 1px solid rgb(230, 230, 230);
+}
