+++ src/components/Divider/Divider.ts
@@ -1,10 +1,10 @@
-import { LitElement, html } from "lit";
+import { LitElement, html, unsafeCSS } from "lit";
 import { customElement } from "lit/decorators.js";
-import DividerStyle from "./DividerStyle";
+import DividerStyle from "./Divider.css" assert { type: "css" };
 
 @customElement("candy-divider")
 export class CandyDivider extends LitElement {
-  static styles = DividerStyle;
+  static styles = unsafeCSS(DividerStyle);
 
   render() {
     return html`<div part="divider" class="divider-container"><div></div></div> `;
