--- src/components/Divider/DividerStyle.ts
@@ -1,12 +1,0 @@
-import { css } from "lit";
-
-export default css`
-  .divider-container {
-    padding: 0.5rem;
-  }
-
-  .divider-container div {
-    display: flex;
-    border-top: 1px solid rgb(230, 230, 230);
-  }
-`;
