+++ src/components/Icon/Icon.css
@@ -1,0 +1,14 @@
+.sm {
+  font-size: 0.75rem;
+  line-height: 1rem;
+}
+
+.md {
+  font-size: 1rem;
+  line-height: 1.5rem;
+}
+
+.xl {
+  font-size: 1.5rem;
+  line-height: 2rem;
+}
