--- src/components/Icon/IconStyle.ts
@@ -1,18 +1,0 @@
-import { css } from "lit";
-
-export default css`
-  .sm {
-    font-size: 0.75rem;
-    line-height: 1rem;
-  }
-
-  .md {
-    font-size: 1rem;
-    line-height: 1.5rem;
-  }
-
-  .xl {
-    font-size: 1.5rem;
-    line-height: 2rem;
-  }
-`;
