+++ src/components/Input/Input.ts
@@ -1,6 +1,6 @@
-import { LitElement, html } from "lit";
+import { LitElement, html, unsafeCSS } from "lit";
 import { customElement, property, query } from "lit/decorators.js";
-import InputStyle from "./InputStyle";
+import InputStyle from "./Input.css" assert { type: "css" };
 
 export type CandyInputProps = {
   label: string;
@@ -11,7 +11,7 @@ export type CandyInputProps = {
 
 @customElement("candy-input")
 export class CandyInput extends LitElement {
-  static styles = InputStyle;
+  static styles = unsafeCSS(InputStyle);
 
   @query("input")
   protected _input!: HTMLInputElement;
