+++ src/components/Link/Link.css
@@ -1,0 +1,5 @@
+.link-container {
+  font-weight: 500;
+  color: inherit;
+  text-decoration-line: underline;
+}
