--- src/components/Link/LinkStyle.ts
@@ -1,9 +1,0 @@
-import { css } from "lit";
-
-export default css`
-  .link-container {
-    font-weight: 500;
-    color: inherit;
-    text-decoration-line: underline;
-  }
-`;
