+++ src/components/SidebarDoc/DocElement/DocElement.ts
@@ -1,6 +1,7 @@
-import { LitElement, html } from "lit";
+import { LitElement, html, unsafeCSS } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import ElementStyle from "./DocElementStyle";
+
+import ElementStyle from "./DocElement.css" assert { type: "css" };
 import Icon from "../../../assets/icons/candy-icon.svg";
 
 export type CandySidebarDocElementProps = {
@@ -11,7 +12,7 @@ export type CandySidebarDocElementProps = {
 
 @customElement("candy-sidebar-doc-element")
 export class CandySidebarDocElement extends LitElement {
-  static styles = ElementStyle;
+  static styles = unsafeCSS(ElementStyle);
 
   @property({ type: String })
   label = "";
